requirements.txt
mega.py==1.0.8
python-dotenv==1.0.0
